Testing JUCE with Cmake.
It's time to start making some vst's, and The Audio Programmer on YouTube seems to have some good resources for working with the JUCE framework with Cmake. So TestJuceCmakePlugin is me getting a JUCE plugin project to build and run using Cmake, and it worked pretty well.

Note: JUCE is not compatible with MinGW so remember to use the MSVS toolchain.